cloud/uris: break dependency on cloud/cloudpb
We need to break the dependency of cloud/uris on cloud/cloudpb to
satisfy various disallowed import constraints in sql/sem/tree when we
import cloud/uris from there. Fortunately this is straightforward: we
just have to move a few functions from uris.go to external_storage.go.

Epic: None

Release note: None
