+++ pkg/cloud/cloudpb/external_storage.go
@@ -95,3 +95,41 @@ func DecodeRoleProviderString(roleProviderString string) (p ExternalStorage_Assu
 	}
 	return p
 }
+
+// ParseRoleString parses a comma separated string of roles into a list of
+// intermediate delegate roles and the final assumed role.
+func ParseRoleString(roleString string) (string, []string) {
+	roleProvider, delegateRoleProviders := ParseRoleProvidersString(roleString)
+
+	delegateRoles := make([]string, len(delegateRoleProviders))
+	for i := range delegateRoleProviders {
+		delegateRoles[i] = delegateRoleProviders[i].Role
+	}
+	return roleProvider.Role, delegateRoles
+}
+
+// ParseRoleProvidersString parses a comma separated list of role provider
+// strings.
+//
+// Each role provider string is in the format: "<role>;external_id=<id>"
+// The external ID portion of the format, including the ';', is optional and can
+// be omitted if there is no external ID needed when assuming the role.
+func ParseRoleProvidersString(
+	roleProvidersString string,
+) (
+	assumeRole ExternalStorage_AssumeRoleProvider,
+	delegateRoles []ExternalStorage_AssumeRoleProvider,
+) {
+	if roleProvidersString == "" {
+		return assumeRole, delegateRoles
+	}
+
+	roleProviders := strings.Split(roleProvidersString, ",")
+	delegateRoles = make([]ExternalStorage_AssumeRoleProvider, len(roleProviders)-1)
+
+	assumeRole = DecodeRoleProviderString(roleProviders[len(roleProviders)-1])
+	for i := 0; i < len(roleProviders)-1; i++ {
+		delegateRoles[i] = DecodeRoleProviderString(roleProviders[i])
+	}
+	return assumeRole, delegateRoles
+}
